requirements.txt
requests
beautifulsoup4
configparser
psycopg2-binary
pandas
pytest